pyproject.toml
[project]
name = "cadri-cli"
version = "0.1.0"
description = "CLI for AWS EC2 image preparation, launch, and result collection."
requires-python = ">=3.11"
dependencies = [
    "argcomplete>=3.6.3",
    "boto3>=1.34",
    "PyYAML>=6.0",
]

[project.scripts]
cadri = "cadri_cli.cli:main"

[tool.setuptools.packages.find]
include = ["cadri_cli*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
